+++ tests/testthat/test-transformations.R
@@ -7,12 +7,20 @@ test_that("Testing transformations of data", {
   
   # Perform log transformation
   taxa$scale(transform = log2)
-  expect_snapshot(taxa)
-  
-  # In between reset
-  taxa$reset()
+  expect_snapshot(as.vector(taxa$countData[, 1]))
   
   # Perform sqrt transformation
+  taxa$reset()
   taxa$scale(transform = sqrt)
-  expect_snapshot(taxa)
+  expect_snapshot(as.vector(taxa$countData[, 1]))
+  
+  # Perform clr standardisation
+  taxa$reset()
+  taxa$scale(method = "clr")
+  expect_snapshot(as.vector(taxa$countData[, 1]))
+  
+  # Perform tss normalisation
+  taxa$reset()
+  taxa$scale(method = "tss")
+  expect_snapshot(as.vector(taxa$countData[, 1]))
 })