+++ CHANGELOG.md
@@ -5,12 +5,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v1.5.2 - [2026-04-]
 
+### `Added`
+- [#34](https://github.com/agusinac/OmicFlow/issues/34) Extra argument `group_by` in `foldchange`.
+
 ### `Changed`
 - [#33](https://github.com/agusinac/OmicFlow/issues/33) Combined two methods into a single `omics$scale()`, added euclidean method and aitchison
+- [#36](https://github.com/agusinac/OmicFlow/issues/36) Created `foldchange` method in `omics` abstract class and in `metagenomics` subclass
 
 ### `Fixed`
 - [#35](https://github.com/agusinac/OmicFlow/issues/35) changed removezeros by using `@p` of `CsparseMatrix`
 - [#37](https://github.com/agusinac/OmicFlow/issues/37) `omics$reset` works with `omics$copy`
+
+### `Deprecated`
+- [#36](https://github.com/agusinac/OmicFlow/issues/36) Moved `foldchange` from namespace to `omics$foldchange` and `metagenomics$foldchange`
 
 ## v1.5.1 - [2026-02-27]